Fixed bugs in Commerce class

+++ Market_App/Models/Product.cs
@@ -8,9 +8,9 @@ namespace Market_App.Models
     {
         public int ID { get; set; }
         public string Name { get; set; }
-        public int Price { get; set; }
+        public double Price { get; set; }
         public string Unit { get; set; }
-        public int Residue { get; set; }
+        public float Residue { get; set; }
         public string Type { get; set; }
         public float Purchased { get; set; }
     }
